foobar2000-shellext: Creating a symlink in %AppData%
In version < 2.0, the default location for storing user data was $dir/profile,
even if the file $dir/portable_mode_enabled did not exist. However, in version
2.0, this default path was changed to $env:AppData/foobar2000-v2.

To resolve this issue, we first create a profile folder in $persist_dir
and then create a symlink in $AppData/foobar2000-v2. The existing user data
will be migrated accordingly.

This closes #3.